+++ R/plume-package.R
@@ -15,7 +15,7 @@
 #' @importFrom readr read_file write_lines
 #' @importFrom rlang %||% := abort set_names
 #' @importFrom rlang enexprs sym syms
-#' @importFrom rlang is_named is_string is_bool is_true
+#' @importFrom rlang is_named have_name is_string is_bool is_true
 #' @importFrom rlang caller_env current_env caller_arg
 #' @importFrom glue glue glue_collapse
 #' @importFrom vctrs vec_group_id vec_duplicate_any vec_restore vec_rank